requirements.txt
asgiref==3.2.10
attrs==20.3.0
Django==3.0.8
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.2
